--- src/editor/groups.cpp
@@ -1,6 +1,0 @@
-#include "groups.h"
-
-
-void Groups::onGUI()
-{
-}
