--- src/editor/groups.h
@@ -1,16 +1,0 @@
-#pragma once
-
-
-#include "core/array.h"
-#include "core/string.h"
-
-
-class Groups
-{
-public:
-	void onGUI();
-
-private:
-	Lumix::Array<Lumix::string> m_group_names;
-	Lumix::Array<Lumix::Array<Lumix::Entity> > m_groups;
-};
