+++ src/renderer/editor/terrain_editor.cpp
@@ -21,7 +21,6 @@
 #include "engine/property_register.h"
 #include "renderer/material.h"
 #include "renderer/model.h"
-#include "renderer/ray_cast_model_hit.h"
 #include "renderer/render_scene.h"
 #include "renderer/texture.h"
 #include "renderer/terrain.h"
