+++ src/renderer/terrain.cpp
@@ -12,7 +12,6 @@
 #include "engine.h"
 #include "renderer/material.h"
 #include "renderer/model.h"
-#include "renderer/ray_cast_model_hit.h"
 #include "renderer/render_scene.h"
 #include "renderer/shader.h"
 #include "renderer/texture.h"
